refactor(web): extract Header, Nav, Footer from Layout
Split Layout.tsx into separate component files:
- Header.tsx (with private UserMenu helper)
- Nav.tsx
- Footer.tsx
- Layout.tsx is now a thin composition wrapper

Closes #575
