+++ apps/web/src/components/Footer.tsx
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2026 Alex Brandt <alunduil@gmail.com>
+// SPDX-License-Identifier: MIT
+
+export function Footer() {
+  return (
+    <footer className="border-t border-border bg-muted py-8">
+      <div className="mx-auto max-w-7xl px-4 text-center text-sm text-muted-foreground">
+        <p>Built with React 19 + TypeScript + Vite</p>
+        <p className="mt-2 text-xs text-muted-foreground/70">© 2026 Genshin Team Builder</p>
+      </div>
+    </footer>
+  );
+}
